# perf: optimize getting virtual machine errors