### What I did

Before:
```
Wall time: 554 ms
```

Then with https://github.com/ApeWorX/ape/pull/2211

```
Wall time: 450 ms
```

Then with this PR and not https://github.com/ApeWorX/ape/pull/2211

```
Wall time: 273 ms
```

Then with this PR AND https://github.com/ApeWorX/ape/pull/2211

```
Wall time: 245 ms
```

### How I did it

I think the main thing was not defining such a large method within another method
but also some condition combines probably helped.

### How to verify it

should work like it did before

### Checklist

- [ ] Passes all linting checks (pre-commit and CI jobs)
- [ ] New test cases have been added and are passing
- [ ] Documentation has been updated
- [ ] PR title follows [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) standard (will be automatically included in the changelog)